twitter-stream/pyproject.toml
[tool.poetry]
name = "py-twitter-stream"
version = "0.1.0"
description = ""
authors = ["Alan Bixby <alanrbixby@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.8.3"
pymongo = "^4.3.2"
asyncio = "^3.4.3"
aiodns = "^3.0.0"
decouple = "^0.0.7"
motor = "^3.1.1"
python-dateutil = "^2.8.2"
python-dotenv = "^0.21.0"

[tool.poetry.group.dev.dependencies]
mypy = "^0.982"
flake8 = "^5.0.4"
black = {version = "^22.10.0", allow-prereleases = true}
types-python-dateutil = "^2.8.19.2"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
